webhook/functions/requirements.txt
supabase
google-cloud-documentai
python-dotenv
langchain
pydantic
openai
langchain-community